'Development': stream docker container logs to NATS subject (#221)
* Switch from Redis to NATS  (#2)

* Refactor job queue handling to support priority-based processing and update related tests

* Enhance job stream configuration to disallow duplicates

* Add KeyValue store for job management

* Add Bruno files for NATS

* stream container logs to NATS subject

* Replace Redis configuration with NATS options in .env.example

* Refactor configuration to replace Redis with NATS

* Update spf13/cast dependency to v1.7.1 in go.work.sum

* `Development`: Fix missing shared folder in volume between containers (#219)

* update bruno script to show failure if it fails

* add shared volumes to fix "folder not found error"

* ensure debug mode logs deletion of volume

* implement delete volume after creation before execution flag

* Implement Rabbit Feedback

* Remove wrong usage of endpoint() in test

* update log structure

* add manual removeContainer to allow log processing

* log structure abstraction, initial k8s extractor

* remove k8s log extraction, add log structure abstraction

* revert k8s changes

* go mod tidy after conflict resolution

* add nats connection nil check in PublishLogsToNATS()

* update parser

* publisher interface, remove nats dependency (nc)

* rabbit feedback

* refactor HadesScheduler docker file structure

* implement rabbit feedback

* add scheduler fail handling

* implement rabbit feedback

* remove redundant remove container

---------
