+++ .env.example
@@ -22,4 +22,4 @@ DOCKER_HOST="unix:///var/run/docker.sock" # Docker host to connect to (default: 
 DOCKER_SCRIPT_EXECUTOR="/bin/bash -c" # Path to the script executor to use (default: /bin/sh)
 DOCKER_CPU_LIMIT=2 # Number of CPUs to limit the container to (see Docker documentation for more info)
 DOCKER_MEMORY_LIMIT=2g # Memory limit for the container (see Docker documentation for more info)
-DOCKER_CONTAINER_AUTOREMOVE=true # Automatically remove containers after they finish running (default: true)
+DOCKER_CONTAINER_AUTOREMOVE=false # Disable auto-removal to retain container logs post-execution (default: false)
